Feature: Add tab(s) to existing session(s)
Add a set of tab(s) to existing session(s).
Choose the set of tabs and select the session
to add them to. If multiple existing-sessions
are chosen, the tabs will be added to all those
sessions. Verified basic functionality to be
working.

The existing query-calls have been modified to
reduce code-churn due to the feature.